+++ packages/scroll-carousel/test/fixtures/astro/astro.config.mjs
@@ -1,0 +1,7 @@
+import { defineConfig } from 'astro/config';
+
+export default defineConfig({
+  base: '/scroll-carousel/fixtures/astro',
+  outDir: '../../../../../_site/scroll-carousel/fixtures/astro',
+  logLevel: 'warn',
+});