+++ test/fixtures/astro/astro.config.mjs
@@ -1,7 +1,0 @@
-import { defineConfig } from 'astro/config';
-
-export default defineConfig({
-  base: '/fixtures/astro',
-  outDir: '../../../_site/fixtures/astro',
-  logLevel: 'warn',
-});